Allow running scripts when building MOTD
Currently we can have multiple files in /etc/motd.d/ and their content
would be appended to MOTD. This change would check if the file is
executable and, if so, append the ouput of the file to MOTD, not the
content of the file. Non executable files would still get their content
appended to MOTD. This also preserves alphabetical ordering.